requirements.txt
requests==2.25.1
urllib3==1.26.4
python-dotenv==0.15.0
flake8==3.9.0
black==20.8b1